add flag for read only mode

+++ schafkopf/core/dynamodb/email_table.py
@@ -11,10 +11,16 @@ class EmailItem(BaseModel):
 
 
 def add(dynamodb, email: EmailItem):
+    if not env.on_aws():
+        print(f"Not on AWS, not adding: {email}")
+        return
     table = dynamodb.Table("schafkopf_emails")
     table.put_item(Item=json.loads(email.model_dump_json()))
 
 def delete(dynamodb, email: str):
+    if env.read_only():
+        print(f"Read only, not deleting: {email}")
+        return
     table = dynamodb.Table("schafkopf_emails")
     table.delete_item(Key={"email": email})
